+++ payloads/v9/_interactions/base.ts
@@ -51,8 +51,7 @@ export interface APIBaseInteractionMetadata<Type extends InteractionType> {
 /**
  * @see {@link https://discord.com/developers/docs/resources/message#message-interaction-metadata-object-application-command-interaction-metadata-structure}
  */
-export interface APIApplicationCommandInteractionMetadata
-	extends APIBaseInteractionMetadata<InteractionType.ApplicationCommand> {
+export interface APIApplicationCommandInteractionMetadata extends APIBaseInteractionMetadata<InteractionType.ApplicationCommand> {
 	/**
 	 * The user the command was run on, present only on user commands interactions
 	 */
@@ -67,8 +66,7 @@ export interface APIApplicationCommandInteractionMetadata
 /**
  * @see {@link https://discord.com/developers/docs/resources/message#message-interaction-metadata-object-message-command-interaction-metadata-structure}
  */
-export interface APIMessageComponentInteractionMetadata
-	extends APIBaseInteractionMetadata<InteractionType.MessageComponent> {
+export interface APIMessageComponentInteractionMetadata extends APIBaseInteractionMetadata<InteractionType.MessageComponent> {
 	/**
 	 * ID of the message that contained the interactive component
 	 */
@@ -257,10 +255,7 @@ export type APIInteractionDataResolvedChannel =
  * @see {@link https://discord.com/developers/docs/resources/guild#guild-member-object}
  */
 export interface APIInteractionDataResolvedGuildMember
-	extends APIBaseGuildMember,
-		APIFlaggedGuildMember,
-		APIGuildMemberAvatar,
-		APIGuildMemberJoined {
+	extends APIBaseGuildMember, APIFlaggedGuildMember, APIGuildMemberAvatar, APIGuildMemberJoined {
 	permissions: Permissions;
 }